test: add redis open check on test

+++ tests/fixtures/setup-db.ts
@@ -49,7 +49,9 @@ export const resetDatabase = async () => {
     await RosterRequest.deleteMany({})
     await ArchiveTeam.deleteMany({})
     await OneTimePasscode.deleteMany({})
-    await redisClient.flushAll()
+    if (redisClient.isOpen) {
+        await redisClient.flushAll()
+    }
 }
 
 export const tearDownDatabase = () => {
